requirements_gpu.txt
addict==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
av==11.0.0
babel==2.17.0
bitarray==3.5.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
Cython==3.1.2
decorator==5.2.1
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
espeakng-loader==0.2.4
exceptiongroup==1.3.0
fairseq==0.12.2
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
h11==0.16.0
hatesonar-onnxruntime==0.0.8
hf-xet==1.1.5
huggingface-hub==0.33.4
humanfriendly==10.0
hydra-core==1.0.7
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kokoro-onnx==0.4.9
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
misaki==0.9.4
mpmath==1.3.0
msgpack==1.1.1
murmurhash==1.0.13
networkx==3.4.2
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.0.6
onnxruntime-gpu==1.22.0
packaging==25.0
pastel==0.2.1
phonemizer-fork==3.3.2
platformdirs==4.3.8
poethepoet==0.24.4
pooch==1.8.2
portalocker==3.2.0
praat-parselmouth==0.4.6
preshed==3.0.10
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.6
pyworld==0.3.5
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
resampy==0.4.3
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.26.0
rvc==0.3.4
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
segments==2.3.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.1
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.4
tensorrt==10.12.0.36
tensorrt-cu12==10.12.0.36
tensorrt_cu12_bindings==10.12.0.36
tensorrt_cu12_libs==10.12.0.36
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchcrepe==0.0.22
tqdm==4.67.1
transformers==4.53.1
triton==3.1.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
